+++ src/main/java/com/haejwo/tripcometrue/domain/triprecord/repository/triprecord_schedule/TripRecordScheduleRepository.java
@@ -3,6 +3,8 @@ package com.haejwo.tripcometrue.domain.triprecord.repository.triprecord_schedule
 import com.haejwo.tripcometrue.domain.triprecord.entity.TripRecordSchedule;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TripRecordScheduleRepository extends JpaRepository<TripRecordSchedule, Long> {
+public interface TripRecordScheduleRepository extends
+    JpaRepository<TripRecordSchedule, Long>,
+    TripRecordScheduleCustomRepository {
 
 }
